analytics_service/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-pubsub
python-dotenv
